+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 21:57:48 by dpoveda-          #+#    #+#             */
-/*   Updated: 2021/11/09 22:17:37 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2021/11/10 09:39:25 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,7 @@
 
 # include <libft.h>
 # include <unistd.h>
+
+int	sh_loop();
 
 #endif